// src/components/Game/components/Board/types.ts
export type BoardCellCoordinates = {
  rowIndex: number;
  cellIndex: number;
};

export type BoardClientPoint = {
  clientX: number;
  clientY: number;
};

export type BoardClientRect = {
  top: number;
  left: number;
  width: number;
  height: number;
  bottom: number;
  right: number;
};

export type BoardGridSize = {
  rows: number;
  cols: number;
};